+++ cmp/grid/filter/GridFilterModel.js
@@ -23,7 +23,7 @@ export class GridFilterModel extends HoistModel {
     /** @member {(Store|View)} */
     bind;
     /** @member {boolean} */
-    autoApply;
+    commitOnChange;
     /** @member {GridFilterFieldSpec[]} */
     @managed fieldSpecs = [];
 
@@ -41,8 +41,7 @@ export class GridFilterModel extends HoistModel {
      * @param {(Store|View)} c.bind - Store or cube View that should actually be filtered
      *      as column filters are applied, and used to provide suggested data values in
      *      column filters (if configured).
-     * @param {boolean} [c.autoApply] - true (default) to automatically apply filters to the
-     *      bound Store / View.
+     * @param {boolean} [c.commitOnChange] - true (default) to eagerly commit filters on each change.
      * @param {(string[]|Object[]} [c.fieldSpecs] - specifies the fields this model supports
      *      for filtering. Should be configs for a `GridFilterFieldSpec`. These may be specified
      *      as field names in bound Store/View or omitted entirely, indicating that all fields
@@ -53,7 +52,7 @@ export class GridFilterModel extends HoistModel {
     constructor({
         gridModel,
         bind,
-        autoApply = true,
+        commitOnChange = true,
         fieldSpecs,
         fieldSpecDefaults
     }) {
@@ -61,7 +60,7 @@ export class GridFilterModel extends HoistModel {
         makeObservable(this);
         this.gridModel = gridModel;
         this.bind = bind;
-        this.autoApply = autoApply;
+        this.commitOnChange = commitOnChange;
         this.fieldSpecs = this.parseFieldSpecs(fieldSpecs, fieldSpecDefaults);
     }